+++ graphique/scalars.py
@@ -3,10 +3,10 @@ GraphQL scalars.
 """
 
 import functools
-import re
 from datetime import date, datetime, time, timedelta
 from decimal import Decimal
-from typing import Union, no_type_check
+from typing import Union
+import isodate
 import pyarrow as pa
 import strawberry
 
@@ -17,47 +17,23 @@ def parse_long(value) -> int:
     raise TypeError(f"Long cannot represent value: {value}")
 
 
-@no_type_check
-def parse_duration(value: str):
-    months = days = seconds = 0
-    d_val, _, t_val = value.partition('T')
-    parts = re.split(r'(-?\d+\.?\d*)', d_val.lower() + t_val)
-    if parts.pop(0) != 'p':
-        raise ValueError("Duration format must start with `P`")
-    multipliers = {'y': 12, 'w': 7, 'H': 3600, 'M': 60}
-    for num, key in zip(parts[::2], parts[1::2]):
-        value = (float if '.' in num else int)(num) * multipliers.get(key, 1)
-        if key in 'ym':
-            months += value
-        elif key in 'wd':
-            days += value
-        elif key in 'HMS':
-            seconds += value
-        else:
-            raise ValueError(f"Invalid duration field: {key.upper()}")
-    if set(d_val).isdisjoint('YM'):
-        return timedelta(days, seconds)
-    return pa.MonthDayNano([months, days, int(seconds * 1_000_000_000)])
+def parse_duration(value):
+    duration = isodate.parse_duration(value)
+    if isinstance(duration, timedelta) and set(value.partition('T')[0]).isdisjoint('YM'):
+        return duration
+    months = getattr(duration, 'years', 0) * 12 + getattr(duration, 'months', 0)
+    nanoseconds = duration.seconds * 1_000_000_000 + duration.microseconds * 1_000
+    return pa.MonthDayNano([months, duration.days, nanoseconds])
 
 
-@functools.singledispatch
-def duration_isoformat(months: int, days: int, seconds: int, fraction: str = '.') -> str:
-    minutes, seconds = divmod(seconds, 60)
-    items = zip('YMDHM', divmod(months, 12) + (days,) + divmod(minutes, 60))
-    year, month, day, hour, minute = (f'{value}{key}' if value else '' for key, value in items)
-    fraction = fraction.rstrip('0').rstrip('.')
-    return f'P{year}{month}{day}T{hour}{minute}{seconds}{fraction}S'
-
-
-@duration_isoformat.register
-def _(td: timedelta) -> str:  # type: ignore
-    return duration_isoformat(0, td.days, td.seconds, f'.{td.microseconds:06}')
+duration_isoformat = functools.singledispatch(isodate.duration_isoformat)
 
 
 @duration_isoformat.register
 def _(mdn: pa.MonthDayNano) -> str:
-    seconds, nanoseconds = divmod(mdn.nanoseconds, 1_000_000_000)
-    value = duration_isoformat(mdn.months, mdn.days, seconds, f'.{nanoseconds:09}')
+    value = isodate.duration_isoformat(
+        isodate.Duration(months=mdn.months, days=mdn.days, microseconds=mdn.nanoseconds // 1_000)
+    )
     return value if mdn.months else value.replace('P', 'P0M')